+++ CHANGELOG.md
@@ -4,9 +4,16 @@
 
 ### Added
 
+- Add `mypy` as a lint dependency. Also publish type information for when `stactools`
+  is used as a library. ([#182](https://github.com/stac-utils/stactools/pull/182))
 - Optional dependency on s3fs ([#178](https://github.com/stac-utils/stactools/pull/178)), enabling:
   - Using s3 files as external data for testing
   - Using s3 hrefs with stactools functionality by installing with `pip install stactools[s3]` (or `pip install stactools[all]`)
+- `stac validate` command for validating JSON and checking links ([#151](https://github.com/stac-utils/stactools/pull/151))
+
+### Changed
+
+- Updated PySTAC dependency to `~= 1.1` ([#185](https://github.com/stac-utils/stactools/pull/185))
 
 ## stactools 0.2.1a2